git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@6009 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/ps.c
@@ -671,7 +671,7 @@ static Image *ReadPSImage(const ImageInfo *image_info,ExceptionInfo *exception)
           &bounds.x1,&bounds.y1,&bounds.x2,&bounds.y2);
         i=1;
       }
-    if ((count != 4) || (i < priority))
+    if ((count != 4) || (i < (ssize_t) priority))
       continue;
     hires_bounds=bounds;
     priority=i;
